Split DSS cli/api tests into 2 files.
